Fixes #15
This was an easy one, especially, thanks to @dictino comments. The only required change is in JSONRPC-BuildInfo.vi

Other than that, I made more clear in TestOK.vi the possibility of using hidden indicators and controls that are not exchanged with a RIP client. I also changed the name of the testing vis to remove the reference to the old JiL and added a description to make more clear their purpose.

The rest of commited files are... well, you know how LabVIEW works...